persist & load settings from disk
this is pretty rudimentary, but it works for WASM & desktop which is
good enough for now!

quad-storage works out of the box, but it has some limitations like the
location is always set to `local.data` on desktop.

in a more ideal world, I think the data structure would be verified and
backed by a bit more